requirements.txt
anyio==4.9.0
blinker==1.9.0
boto3==1.37.18
botocore==1.37.18
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.44.0
dnspython==2.7.0
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.25.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
pymongo==4.11.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==20.7
requests==2.32.3
requests-toolbelt==0.10.1
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==1.26.15
Werkzeug==3.1.3
